Moved to lib/
Moved source files into the expected directory, lib. This means that .inch.yml and yardopts can be deleted - as this is now the default configuration anyway.
