--- .inch.yml
@@ -1,4 +1,0 @@
-files:
-    included:
-        - /*.rb
-        - /*.md
