+++ tests/getfunctions.rb
@@ -1,4 +1,4 @@
-require_relative "../main.rb"
+require_relative "../lib/main.rb"
 
 require "coveralls"
 require "codacy-coverage"
