--- yardopts
@@ -1,2 +1,0 @@
-/*.md
-/*.rb
